Implemented xdebug profiling
Summary: Collecting profiling was implemented in D1423763, enabling/disabling it was
implemented in D1438290, this implements writing the collected profiling info
to a cachegrind compatible file.

The cachegrind file (at least the xdebug version) describes function calls,
function call files, function call time, and function call line #'s.

Programs that read this format (such as kcachegrind, qcachegrind, etc) can
use the generated files to get pretty graphs:
https://www.facebook.com/pxlcld/lqww
https://www.facebook.com/pxlcld/lqwz

One major difference between php5 xdebug and this is that only certain builtin
functions will be collected by the profiler. This is very unfortunate as
profiling builtin function calls would definitely be useful to the user, but
there isn't an obvious fix to this. The profilers plug into
the runtime via event hooks, which are not called for builtin function begin/end
events. Adding this wouldn't be simple because the event hook function begin/end
methods assume an ActRec is generated, which is not true for builtins.
Depends on D1438290.

Reviewed By: @ptarjan

Differential Revision: D1448800