Fixed xdebug tracing tests
Summary: This diff fixes some of the xdebug tracing tests. When tracing turned on via an
ini setting, the filename that is used is by default non-unique. In the tests
related to tracing ini settings, I mistakingly didn't change the file output
name from default.

This was discovered when the tracing_2.php test failed in D20697. It interfered
with tracing_basic_1.php.

Reviewed By: @JoelMarcey

Differential Revision: D1460860