Only track live objects that actually have destructors
Summary: When Eval.EnableObjDestructCall is on, we have to keep a set of live
objects to destruct at the end of the request. In practice, most objects don't
have user-defined destructors, so short-circuit the slow std::set operations
when they're not necessary.

I wrote a little microbenchmark that does nothing but create a bunch of objects
kept alive by cyclic references; with this diff it runs in 42% of the time it
used to:

  <?php
  class cycle {
    private $obj;
    public function __construct() { $this->obj = $this; }
  }
  for ($i = 0; $i < 5000000; ++$i) new cycle;

Reviewed By: @jdelong

Differential Revision: D1462739