Sanitize non-utf8 content in test output
Summary: Ran into this with a test which outputs cp1251 text
which json_encode (rightly) barfs on.  This results in the
test's diff output being reduced to null which is singularly
unhelpful in test output.

Use UConverter::transcode() to turn anything that isn't UTF-8
into the Unicode WTF character U+FFFD.

Hopefully we don't get too many of these kinds of tests.

Reviewed By: @ptarjan

Differential Revision: D1462059