added a new Config::Bind that works for ini settings
Summary: There was no good way to variables to ini settings that did not automatically enable them for config.hdf. Now there is a Config::Bind function that will bind an ini setting to a variable using the value specified in the ini file or the a given default otherwise. The main difference between this Config::Bind and the old/other one is that it does not prepend "hhvm." on the front and does not override the ini setting with a hdf setting.

D1474476 depends on this diff and had a clean run. Given that this diff runs no new code it seems safe to override failures.

Reviewed By: @ptarjan

Differential Revision: D1463584