Turn on Eval.EnableObjDestructCall by default
Summary: This was an intentional divergence from PHP5 for performance reasons,
but the default should be compatible. The overhead of this feature is vastly
reduced with D1462739 anyway.

This change by itself would've reduced test coverage of fast object allocaton
paths in the jit, but instead of working around that I just properly
implemented support for EnableObjDestructCall being true. I also removed a
bogus check for it in emitCreateCl. Finally, it exposed a bug in
Class::initSPropHandles where we were incorrectly marking some class's SProp
handles as persistent even if the parent's SProp handle wasn't persistent (this
was causing test/quick/properties4.php to fail).

Reviewed By: @alexmalyshev

Differential Revision: D1466192