Don't smart-allocate anything when creating the execution context
Summary: We (re-)create the execution context at the end of every
request. If it smart allocates anything, we will end up allocating
a 2Mb slab which won't be freed (or used) until the next time
the thread executes a request. Since we typically have a lot of
idle requests, this could be a significant chunk of memory.

Reviewed By: @bertmaher

Differential Revision: D1465987