Fix race condition in DebuggerProxy start/stop
Summary: hphpd is sometimes crashing while calling pollSignal() on an already
destructed DebuggerProxy from an AsyncFunc.  (This manifests as a pure
virtual call to ThriftBuffer::flushImpl()).

It looks like there's a potential race condition where we might stop the
proxy and clean it up after we've added it to the Debugger's map, but
before we've actually started it up.  Let's fix this and hope it fixes
the crash also.

Reviewed By: @markw65

Differential Revision: D1468019