Support "return;" in generators
Summary: When FB originally built generators for HPHPc way back in the
day, we used `yield break;` to exit from a generator. When PHP picked up
the feature, they just used `return;` -- so return with no value is
legal in a generator (but returning a value is of course not). The
typechecker should understand this behavior.

Reviewed By: @jano

Differential Revision: D1467566