Fix IRBuilder::startBlock
Summary: This function was merging the current FrameState with the block that was about to be started, which is only correct when translating traces. This function was also doing other things that didn't seem necessary (clearing CSE stuff and compatibility checks), which I removed. Finally, this function was doing some block-chaining that isn't correct in general and only seems to be needed for inlining. So I moved this to translateRegion. The end result is that IRBuilder::startBlock now only saves the state for the current block and the switches to the new block.

This diff also adds to Translator a pointer to the current region being translated, which is useful for debugging. This helped me find a bug in createBlockMaps that was causing an assert failure in the refcount validation.

Reviewed By: @bertmaher

Differential Revision: D1436362