Add LibGMP extension
Summary: I've recreated the PHP gmp extension with HNI.  I believe it to be working well, but this is the first extension I've written so any and all criticism of my C-code is encouraged.

This extensions supports all GMP functions, and constants, with the exception of gmp_random.

I have submitted a CLA in the past; it might be under "ian@ my github username" (CA -> .ca); I had to submit it a few times last time around, so hopefully it's properly on record this time; if it's not, let me know.

Cheers

Closes https://github.com/facebook/hhvm/pull/3314

Reviewed By: @ptarjan

Differential Revision: D1470423

Pulled By: @sgolemon