Print current approx typechecking position on SIGUSR1
Summary: When some code being typechecked causes pathological behavior
in the typechecker, it can be useful to figure out approximately where
in the codebase we are checking. Right now, this involves a lot of
one-off logging being manually added, so let's do something a little
more clever. Every time we typecheck an expression, record its position,
and then on SIGUSR1, print that out. The envisoned use case is to attach
to a runaway proccess with gdb, look at the backtrace, and kill it with
SIGUSR1 to try and figure out WTF is going on.

We may ultimately want to expand this into something more general,
notably including things other than typing, but this is good enough for
now -- the 3-4 cases of runaway workers I've seen have all been in
typing.

This also kills `Typing.debug` since it was only used in one place, was
potentially confusing with the new data I'm adding, and was of dubious
utility (blocking `hh_show` in server mode).

Reviewed By: @elgenie

Differential Revision: D1467911