Fix usage of field_key/value in yield
Summary: This wasn't morally doing the right thing, even though the
current usages of `field_key` and `field_value` don't let you trigger
it. The type of an array field key and a `yield $x => $y` field key are
different mechanics, and the logic in the latter to determine whether a
missing `$x` is implicitly null or an int doesn't apply to arrays.

Reviewed By: @jano

Differential Revision: D1471174