Imported needed xdebug files
Summary: This diff directly imports some xdebug code in the directory
`runtime/ext/xdebug/php5_xdebug` as well as in
`runtime/ext/xdebug/xdebug_utils.*`.  Right now, this includes the files
for xdebug strings, xdebug xml, and xdebug memory allocation. I've modified
it to compile in c++, I've removed and replaced dependencies, and I've changed
the style enough to get the linter to stop complaining.

I've written TODOs in each of these files to abstract and refactor
these files. This is low-pri, but I do think it's worth being on the list of
things to do for maintainablity purposes. As I'm sure you will see, the code is
hard to follow, it's verbose, and malloc and free are used worryingly liberally.

The threat of creating potential memory leaks due to the required manual
mallocing/freeing when modifying code is probably my biggest worry. I don't think
cleaning these up will be a huge task, as I've ensured they're as modularized as
possible.

Reviewed By: @ptarjan

Differential Revision: D1471865