fixed using []s in ini settings
Summary: Ini setting arrays were not being incremented properly when the push back syntax was used i.e. extensions[] = "foo.so". The folly::dynamic isInt() function does not check if a type can be converted to an it only if it is currently an int. Unfortunately, there does not appear to be a way to test this other than trying to convert the string to an int and catching the error. It appears there are two options, either we catch the exception (what I implemented) or we just insert the new key as an int. The latter seems like a poor choice because when the setting is later Get()ed any user defined numerical offsets will conflict with our internal numbers and one of the two would dissapear.

Reviewed By: @ptarjan

Differential Revision: D1474402