Force OOM check on large smart allocations
Summary: smartEnlist is used to add large smart allocations that don't
fit within the slabs. When using jemalloc, we account for these using
the per-thread stats, but they don't show up in m_usage. This
effectively means if the largest blocks are on that path, until we are
using enough of tracked, slab-based object allocations to go over max
bytes (which could be a long time) jemalloc might have overflowed but
this is invisible to us.

Technically there is still a hole here, since something outside of the
memory manager could have overflowed jemalloc's per-thread accounting,
but this fixes one obvious source. #4836461 would help us seal it
entirely.

Reviewed By: @jasone

Differential Revision: D1475090