Stop doing sketchy guard optimization in pushStLoc
Summary: I feel this optimization has always been a little dodgy, but it hasn't
bitten us until we started trying to jit loops.

The high-level overview is that the loosened guards resulting from
passing DataTypeGeneric here cause breakage when it happens inside a
loop, if the truth value of the condition the loop is guarding changes
inside the loop.

Reviewed By: @swtaarrs

Differential Revision: D1473217