Fix bug in cgLoad
Summary: If the type is a subtype of Null, we early-exit. We already
detect the dead-load situation, and type of {InitNull|Uninit}
should go down the cgLoadTypedValue case, which should correctly
load.

Now that I look more closely I don't think this is a bug,
but the code was misleading.

Reviewed By: @swtaarrs

Differential Revision: D1478816