Checked map-like arrays should not warn for array_merge
Summary: Renumber is only called in the following places:

* array-common.cpp | ArrayCommon::Dequeue | Remove key then renumber (dequeue should already warn/downgrade and is only used by APCLocalArray and NVTWrapper)
* empty-array.cpp | EmptyArray::Merge | Calls MixedArray::Copy(elems) then renumber them (array_merge shouldn't warn)
* type-array.cpp | Array::mergeImpl | If rhs is empty, m_px->renumber() (array_merge shouldn't warn)

In all the places, we don't need to raise a warning, and in fact we don't want
to. This was an oversight by yours truly.

If we do want to raise warnings at a later point (since array_merge does not
support Collections) the proper way to do it would be in f_array_merge, but not
in ::Renumber.

Reviewed By: @elgenie

Differential Revision: D1478885