Move libevent patch to the main hhvm repo rather than third-party
Summary: third-party should be for actual third-party code,
not code that we've written to deal with third-party code.

Also add a comment section to the top of the patch to
describe what the patch if for and how to apply it.

After this lands, I'll remove the patch from the third-party
repo in a separate commit.

Reviewed By: @ptarjan

Differential Revision: D1479360