Create HHVM man page
Summary: This will be the man page for HHVM. Needed short-term by Debian, this will
be the official man page for other distros as well.

Wrote this in Markdown (because troff is obscene) and used a converter to generate
the manpage itself.

I used `ronn`.

```
% sudo gem install ronn
% /opt/local/ruby-1.8.7/bin/ronn hhvm.md
```

This created a `hhvm.1` file.

So I don't have to deal with the insanity of troff directly, I am thinking of adding ronn (https://github.com/rtomayko/ronn) to third-party2, and then we can use the markdown file
for our source of truth for the man page.

Reviewed By: @ptarjan

Differential Revision: D1477486