Updated reactphp framework tests to 0.4.1
Summary: We're still at `100%` passing tests.

Just a heads up that when 0.5.0 is released, we will have trouble testing it in this way since most of the code has been removed from the repo and moved into separate component repos so the test suite in the main repo no longer hits most of the code.
Closes https://github.com/facebook/hhvm/pull/3377

Reviewed By: @ptarjan

Differential Revision: D1488870