Add a simple zend extension DSO to the new directory hphp/test/ext_zend_compat/hello.
Add a simple unit test that loads this DSO using a .ini file containing just:
  hhvm.dynamic_extensions[hello] = ./test/ext_zend_compat/hello/hello.so
and then executes the php functions implemented by the zend extension.

The long term goal of this test mechanism is to expose before/after bugs
in the zend compat layer, especially in regards to configuration settings.

This new test works(*), but the hello.so file is not (yet) built
automatically by the global cmake mechanisms.

(*) the extension has been cooked to ensure there's a modify handler
for one of the configuration settings, since otherwise the zend compat layer
seg faults.

Questions for reviewers:
  (1) this contains 3rd party code shameless cut&pasted from a tutorial at:
    http://devzone.zend.com/303/extension-writing-part-i-introduction-to-php-and-zend/
  as such, does it need to be quarantined, or otherwise labeled?

  (2) Perhaps there is a better place to put this "hello" directory?

  (3) I don't know how to make the global cmake mechanisms do, effectively:
    cd .../hello; $SOMEPATH/hphpize ; cmake . && make
  only if hphpize has been built already.

  (4) I don't know if I will need to have "good" tests and "bad" tests,
  with the later migrating to the former as the zend compat layer is debugged.