Add a new class SystemProfiler and a default null singleton instance.
This class implements hookpoints of interest to a system profiler,
currently hooking file load and PHP errors.  This list may grow.

The new class also supplies a HotProfiler.

The instance of a SystemProfiler may be made and bound during the initialization
phase of a 3rd party DSO.

While I'm at it, plumb in instantiation of the pre-existing experimental MemoProfiler.

Add constructor and virtual destructor to class SystemProfiler, hopefully as needed to shut up flint.

Conflicts:
	hphp/runtime/ext/ext_hotprofiler.cpp
	hphp/runtime/ext/ext_hotprofiler.h