Fix sample profiler output; make it more efficient
Summary: Open-source xhprof's sampling profiler outputs an array of
strings mapping to strings. We were outputting strings mapping to
singleton string => int arrays, and the int was always 1. This brings it
in line with open source xhprof.

I also changed the internal storage from a hash table to a vector of
timestamps and stacks, and delayed the conversion of timestamps to
strings until profiling ends. This should lower the profiling overhead,
and has the added bonus that the samples come out in sorted order with
no extra work (open-source xhprof does this too).

Reviewed By: @ptarjan

Differential Revision: D1484272