Remove an invalid assert in LitstrTable
Summary: This assertion isn't valid here because we're just getting a handle to
the mutex.  This assertion was failing because we lock the table for reading
during emitHHBCVisitor (while GenerateTextHHBC is on), and meanwhile another
thread tries to mergeLitstr.  This is exactly what the mutex guards against,
since it prevents mergeLitstr from happening until GenerateText is done.

Reviewed By: @mxw

Differential Revision: D1487996