Add an `hhx' gdb command for inspecting bytecode
Summary: I wanted to have this while debugging a core, but it didn't exist, so I
made it.  Maybe it will also be useful to people who aren't me.

Currently, immediates are only printed as numbers, but I'm going to add
a lookup.py which does lots of id-to-type lookups.  That should increase
sanity when debugging with cores or on opt builds when all the lookup
methods are inlined.

Everything is memoized because apparently gdb is really slow but Python
is fast(er).

Reviewed By: @edwinsmith

Differential Revision: D1488347