Implemented stating/stopping xdebug remote server
Summary: This diff implements correctly starting/stopping the xdebug remote debug
server, including initializing/deinitializing the dbgp protocol. The xdebug
server supports all the configuration options available in php5 xdebug.

A lot of this code started as the php5 xdebug code and so most of
the flow corresponds 1:1 to php5 xdebug. However, the code was very heavily
translated- all lines required some change of varying degree.
So feel free to bug me about style/structure as my goal is to keep this file
clean.

What's missing is responding to commands. There are quite a few of these.
Before that, I need to refactor the debugging hook code to work with non-hphpd
debuggers.

Reviewed By: @ptarjan

Differential Revision: D1473636