Don't swallow timeouts and OOMs in check_request_surprise()
Summary: Currently, if a pending exception is registered in the surprise flags,
_and_ we have a timeout or OOM registered, we ignore the latter and only
process the former. In the case of OOMs at least, if this happens, we
will never register an OOM again for the rest of the request, regardless
of what the consequence of the thrown pending exception is.

Let's instead re-set the timeout and OOM flags in this case, so that
they'll be processed on the next check.

Reviewed By: @jano

Differential Revision: D1496336