Makes HHVM recognize <<Memoize>> and generates a new param and changes the function name
Summary: This is the first part of making a <<Memoize>> attribute.
Specifically, this recognizes the attribute on a method, generates a param to hold the result, and changes the name of the underlying funciton.
It currently throws an error so people don't accidentally stumble upon it. The next step will be to make this work for methods with zero args

The admarket test failure was not possibly related to this

Reviewed By: @ptarjan

Differential Revision: D1490770