added new raise_error family functions that skip the top frame of the backtrace.
Summary: There were not functions for raising errors skipping the top frame of the backtrace. Ideally, I would have made this an optional argument to the existing functions, however, the current functions are overloaded with formatting strings and thus that is not simple. This is useful for functions raising errors from systemlib code as otherwise the top frame of the stack will not contain file and line information. See D1486927

Reviewed By: @ptarjan

Differential Revision: D1488146