Force hhi root to be set in master before workers fork
Summary: Make sure that the workers all get the same, preset hhi path,
computed by the master before they fork. This could manifest as a worker
trying to re-extract the hhi files to a preexisting location (since the
random seed is shared across all of them) and hitting an exception when
the directory already exists.

It's pretty weird that we didn't run into this immediately when I did
the hhi embedding -- AFAICT this has been wrong since day one. I think
what saved us is the logic that says "if the number of changes is lower
than N, do it all in the master", where N is about 10 -- so most of the
time this didn't matter. (And it looks like during init we do some other
stuff in the master too or something? In any event, this was hard to
trigger.)

Reviewed By: @elgenie

Differential Revision: D1501184