Have FrameState drop state when hitting an unprocessed predecessor.
Summary: Adds a bitvector to FrameState. When FrameState::startBlock is called
with a block that has a predecessor that wasn't previously visited by
this FrameState, then the FrameState resets its state. The entry block
gets marked in IRBuilder::IRBuilder, because we don't call startBlock
when constructing IR in selectRegion.

Also adds a BCMarker field to FrameState::startBlock to match
IRBuilder::startBlock. This helps with building IR, when we can't get
the BCMarker from the block's first instruction.

Reviewed By: @ottoni

Differential Revision: D1499056