Implemented xdebug server jit connecting
Summary: This diff adds an `XDebugHookHandler` that is intended to respond to debugger
hooks. This depends on D1474540. The handler is used used to start the xdebug
server on exceptions if "jit" mode is enabled.

This required exposing the server attaching code. To do this, I've merged the
request local data with the other xdebug globals as thread local members on
`XDebugExtension`. It didn't make sense to have two places for thread local
globals.

Reviewed By: @ptarjan

Differential Revision: D1474688