Added step out to debugger hooks and implemented step_out xdebug command
Summary: This diff implements the xdebug commands step_out. It does so by calling the
implemented function `phpDebuggerStepOut`.

I took a different approach than hphpd, I've outlined the core difference
below, but I can go into more detail if necessary. To a user of
`phpDebuggerStepOut` the two approaches would be identical. And my approach
could be used by hphpd as well.

I took a different approach because the addition of the
`phpDebuggerFuncEntryHook` and `phpDebuggerFuncExitHook` allows my approach to
be more general, less complex, and significantly less code. The cost is an extra
check every function exit when a debugger is attached.

The core difference between my approach and hphds is that my approach does not
try to set a breakpoint on the return address of the current function.
This is easy in the simple standard case, but when dealing with nested VM states
in things like generators, things get a lot more complicated. Hphpd
iterates up the stack trying to guess the sucessor to the return opcode in order
to set a breakpoint. Each iteration cases on the type of opcode. If the opcode
allows fallthrough we set a breakpoint on the next opcode, if it is a control flow
opcode a breakpoint set on the destination.

Instead, in my implementation we can check on function exit if the step-out
command is active and if the stack depth has decreased. If so, we then force the
interpreter and and we break on the next valid opcode.

Each of these explanations is simplified, hphpd especially. You can see
hphpd's implementation in `runtime/debugger/cmd/cmd_flow_control` and
`runtime/debugger/cmd/cmd_out`.

This diff also includes the following misc changes:
* Added separator for each section in the debug hooks. The sections already
  existed but I noticed when adding my changes the file was getting hard to
  follow.
* Added `phpDebuggerContinue`, which should be called by debuggers to continue
  execution. It just short-ciruits other commands if they are running.
* Unified the logic deciding if we should force debugger interrupts into
  a method of `RequestInjectionData` instead of duplicating it in
  `debugger-hooks`.

Reviewed By: @ptarjan

Differential Revision: D1502519