Added next command to debugger hooks and implemented step_over xdebug command
Summary: This implements the debugger hook command `phpDebuggerNext` and the xdebug
command `step_over`. These commands are used to step the current line of code,
stepping over function calls if necessary.

Hphpd "next" is, at a high level, a step in and then a step out if our stack
depth has increased. There is a lot of special casing, however, most of which
deals with resumable functions.

When stepping in, hphpd breaks on the next line, or in resumable functions, on
any exit opcode (return, yield, await). This is not used in the "step in" command
in hphpd and results in weird behavior as we are already on the line with
the exit opcode. However, This is used in the next command to determine which
opcode to step to next.

After breaking from a step in, the next command looks at the current opcode and,
if it is an await opcode, it sets a breakpoint on the opcode that will be
executed after the await. On yields it does either a step out or a step to the
next line, whichever happens first. On returns, it does a step out. In all of
these cases, in my implementation at least, simply letting the
step in continue would have achieved the same result: we would break on the next
opcode not on this line, which would be either be at a lower stack depth or at
or the resumed opcode for await/yield.

So my implementation does a step in and then a step out if our stack level has
increased. We maintain the "flow filter" from "step in" into "step out". This
allows us to ensure we break on the next line.

There is no casing over types of opcodes. This results in much simpler code and
empirically it results in in behavior identical to hphpd. I've tried to test this
thoroughly but it's possible I may have missed an edge case. Regardless, edge
cases could be much more easily integrated into this than if I had just pulled
hphpd's code.

Though it'd require an understanding of the hphpd next command in order to
refactor, this could be used by hphpd.

Reviewed By: @ptarjan

Differential Revision: D1502723